pyproject.toml
[tool.poetry]
name = "bittrade-kraken-websocket"
version = "0.1.0"
description = ""
authors = ["mat <matt@techspace.asia>"]
documentation = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
bittrade-kraken-rest = "^0.9.0"
reactivex = "^4.0.4"
websocket-client = "^1.4.2"
orjson = "^3.8.3"

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"
ipython = "^8.7.0"
rich = "^12.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
